% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_timeline_class.R
\name{calculate_timeline}
\alias{calculate_timeline}
\title{Calculate a timeline with a configured processor}
\usage{
calculate_timeline(processor, data_frame, ...)
}
\arguments{
\item{processor}{A \code{CustomerTimeline} object.}

\item{data_frame}{A data.frame or data.table containing relationship data.}

\item{...}{Reserved for future extensions.}
}
\value{
A data.table with merged periods.
}
\description{
Calculate a timeline with a configured processor
}
