% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerRelationship-package.R
\docType{package}
\name{customerRelationship}
\alias{customerRelationship-package}
\alias{customerRelationship}
\title{customerRelationship: Customer Relationship Timeline Processing}
\description{
Efficiently processes customer relationship data to identify and merge
consecutive periods with minimal gaps using Rcpp for performance and
data.table for scalability.
}
\author{
\strong{Maintainer}: Patrikios \email{Patrikios@users.noreply.github.com}

}
\keyword{internal}
